jitsi-deployment/assets/meeting_portal_app/requirements.txt
fastapi==0.115.12
itsdangerous==2.2.0
jinja2==3.1.6
psycopg[binary]==3.2.1
PyJWT==2.10.1
python-multipart==0.0.20
uvicorn==0.34.2
